requirements.txt
PyMySQL
python-hcl2